Unit tests are reorganized
