+++ practice-app/unit_test_covid_19.py
@@ -37,6 +37,30 @@ def isPlotDataFetchCorrect():
     if not type(json[0]["Cases"])==int:
         return False
     return True
+
+
+def isCoronavirusJson():
+    
+    list = coronavirus_api.coronavirus_summary_search()     
+    
+    if 'Country' not in list[0]:
+        return False
+    else:
+        return True
+
+
+def isJson():
+    
+    list = coronavirus_api.coronavirusCountryLive("France")     
+    if 'Cases' not in list[0]:
+        return False
+    else:
+        if type(list[0]["Cases"]) == int and list[0]["Status"] == 'confirmed':
+            return True
+        else:
+            return False
+
+
 class SimpleTest(unittest.TestCase):
 
     # Returns True or False.
@@ -44,6 +68,8 @@ class SimpleTest(unittest.TestCase):
         self.assertTrue(isCoronavirusByCountryCorrect() == True)
         self.assertTrue(isPlotDataFetchCorrect() == True)
         self.assertTrue(isWorldStatsCorrect() == True)
+        self.assertTrue(isCoronavirusJson() == True)
+        self.assertTrue(isJson() == True) 
 
 if __name__ == '__main__':
     unittest.main()
